+++ src/Node/NodeInterface.php
@@ -151,6 +151,13 @@ interface NodeInterface
     public function getHeight();
 
     /**
+     * Return the number of nodes in a tree.
+     *
+     * @return int
+     */
+    public function getSize();
+
+    /**
      * Accept method for the visitor pattern (see http://en.wikipedia.org/wiki/Visitor_pattern).
      */
     public function accept(Visitor $visitor);